+++ src/LinkDotNet.NCronJob/JobOption.cs
@@ -21,11 +21,6 @@ public sealed class JobOption
     public object? Parameter { get; set; }
 
     /// <summary>
-    /// Gets or sets the <see cref="IsolationLevel"/> of the executed job.
-    /// </summary>
-    public IsolationLevel IsolationLevel { get; set; }
-
-    /// <summary>
     /// Determines whether the cron expression can specify second-level precision.
     /// </summary>
     /// <remarks>
@@ -35,29 +30,4 @@ public sealed class JobOption
     /// that are scheduled to run very frequently.
     /// </remarks>
     public bool EnableSecondPrecision { get; set; }
-}
-
-/// <summary>
-/// Isolation level of the job. This determines how the job is executed.
-/// </summary>
-public enum IsolationLevel
-{
-    /// <summary>
-    /// The job is executed in the same thread as the scheduler. This is the default value.
-    /// </summary>
-    /// <remarks>
-    /// Running jobs in the same thread as the scheduler can lead to performance issues if the synchronous part of the job takes long.
-    /// The scheduler can't guarantee then to schedule new jobs in the given interval.
-    /// </remarks>
-    None,
-
-    /// <summary>
-    /// The job is executed in a new task. This can be helpful if the synchronous part of the job takes long (before any <c>await</c> is called).
-    /// </summary>
-    /// <remarks>
-    /// This will wrap calls to the job in a new task and therefore not block the scheduler.
-    /// This can decrease scalability as the number of tasks increases. Use this option with caution.
-    /// Almost the same can be achieved with a <c>await Task.Yield()</c> at the beginning of the job.
-    /// </remarks>
-    NewTask
 }