+++ src/LinkDotNet.NCronJob/Registry/CronRegistry.cs
@@ -1,31 +1,26 @@
 using System.Collections.Frozen;
-using System.Collections.Immutable;
 
 namespace LinkDotNet.NCronJob;
 
 internal sealed class CronRegistry : IInstantJobRegistry
 {
+    private readonly JobExecutor jobExecutor;
     private readonly FrozenSet<RegistryEntry> cronJobs;
-    private readonly List<RegistryEntry> instantJobs = [];
 
-    public CronRegistry(IEnumerable<RegistryEntry> jobs)
+    public CronRegistry(IEnumerable<RegistryEntry> jobs, JobExecutor jobExecutor)
     {
+        this.jobExecutor = jobExecutor;
         cronJobs = jobs.Where(c => c.CrontabSchedule is not null).ToFrozenSet();
     }
 
     public IReadOnlyCollection<RegistryEntry> GetAllCronJobs() => cronJobs;
 
-    public IReadOnlyCollection<RegistryEntry> GetAllInstantJobsAndClear()
-    {
-        var returnedJobs = instantJobs.ToImmutableArray();
-        instantJobs.Clear();
-        return returnedJobs;
-    }
-
     /// <inheritdoc />
-    public void AddInstantJob<TJob>(object? parameter = null, IsolationLevel level = IsolationLevel.None)
+    public void RunInstantJob<TJob>(object? parameter = null, CancellationToken token = default)
         where TJob : IJob
     {
-        instantJobs.Add(new RegistryEntry(typeof(TJob), new JobExecutionContext(parameter), level, null));
+        var executionContext = new JobExecutionContext(parameter);
+        var run = new RegistryEntry(typeof(TJob), executionContext, null);
+        jobExecutor.RunJob(run, CancellationToken.None);
     }
 }