+++ src/LinkDotNet.NCronJob/Registry/CronRegsitryEntry.cs
@@ -5,5 +5,4 @@ namespace LinkDotNet.NCronJob;
 internal sealed record RegistryEntry(
     Type Type,
     JobExecutionContext Context,
-    IsolationLevel IsolationLevel,
     CrontabSchedule? CrontabSchedule);