+++ src/LinkDotNet.NCronJob/Registry/IInstantJobRegistry.cs
@@ -6,12 +6,12 @@ namespace LinkDotNet.NCronJob;
 public interface IInstantJobRegistry
 {
     /// <summary>
-    /// Adds an instant job to the registry, which gets directly executed. The instance is retrieved from the container.
+    /// Runs an instant job to the registry, which gets directly executed. The instance is retrieved from the container.
     /// <param name="parameter">An optional parameter that is passed down as the <see cref="JobExecutionContext"/> to the job.</param>
-    /// <param name="level">Sets the isolation level, if given, to the job. See <see cref="IsolationLevel"/>.</param>
+    /// <param name="token">An optional token to cancel the job.</param>
     /// </summary>
     /// <remarks>
     /// The contents of <paramref name="parameter" /> are not serialized and deserialized. It is the reference to the passed in object.
     /// </remarks>
-    void AddInstantJob<TJob>(object? parameter = null, IsolationLevel level = IsolationLevel.None) where TJob : IJob;
+    void RunInstantJob<TJob>(object? parameter = null, CancellationToken token = default) where TJob : IJob;
 }