+++ tests/NCronJob.Tests/TimeProviderFactory.cs
@@ -4,8 +4,15 @@ namespace NCronJob.Tests;
 
 internal static class TimeProviderFactory
 {
-    public static ManualTimeProvider GetTimeProvider()
+    public static ManualTimeProvider GetTimeProvider(TimeSpan? advanceTime = null)
     {
-        return new ManualTimeProvider();
+        return new ManualTimeProvider
+        {
+            AutoAdvanceBehavior = new AutoAdvanceBehavior
+            {
+                UtcNowAdvanceAmount = advanceTime ?? TimeSpan.Zero,
+                TimerAutoTriggerCount = 1,
+            },
+        };
     }
 }